refactor: group hooks in a single file, move d.ts to typings
